+++ tests/unit/neptune/new/client/test_project.py
@@ -97,7 +97,7 @@ class TestClientProject(AbstractExperimentTestMixin, unittest.TestCase):
         "neptune.internal.backends.neptune_backend_mock.NeptuneBackendMock.get_int_attribute",
         new=lambda _, _uuid, _type, _path: IntField(value=42, path=path_to_str(_path)),
     )
-    @patch("neptune.internal.operation_processors.read_only_operation_processor.warn_once")
+    @patch("neptune.core.operation_processors.read_only_operation_processor.warn_once")
     def test_read_only_mode(self, warn_once):
         warned_once.clear()
         with init_project(project=self.PROJECT_NAME, mode="read-only") as project: